!!!!WARNING!!!!!!!!DELETED A  LOT-- CLEANED UP PAGE LAYOUTS!
WARNING! I deleted a lot of my pages that we discussed. I also went into the index.js and fixed the routes. verified all is working... including login. Please be warned! I will continue to develop the User.vue page tomorrow as far as getting the products DB integrated, as well as cleaning up the page. It is a mess, and I understand that!